refactor: extract subject-attribute resolution into _SubjectSectionMixin
First cut at decomposing the 5,896-line simple_tools.py. SimpleToolsHandler
is one ~5,750-line class, so the clean split is the mixin pattern the zim
package already uses (ZimOperations = _ContentMixin + _SearchMixin + …),
not module-level function extraction.

Moves the self-contained subject-attribute section path —
_extract_subject_hint / _resolve_section_for_subject /
_maybe_render_subject_section plus the _SUBJECT_HINT_TO_SECTION /
_WEAK_SUBJECT_HINTS tables and the _TABLE_PLACEHOLDER_RE (used only here) —
into openzim_mcp/subject_section.py as _SubjectSectionMixin. Cross-mixin
dependencies (zim_operations, _track, _coerce_content_offset,
_soft_connector_footer) are declared under TYPE_CHECKING, matching
_ContentMixin's convention.

Pure code movement, no behavior change. simple_tools.py 5,896 → 5,623.
Full suite green (2623 passed); mypy/flake8/isort clean.
